Add experimental code to detect FTP
Introduce experimental code which detects if the
service present on an open port is FTP or not.

Class FtpConn needs a fresh socket object right after socket()
to work properly, because it detects a greeting (ftp server sends
1st message).

Other classes could be constructed in a similar manner, though
the interface stil requires some thought.